test: move testing with real service to integration test
