+++ tests/integration/network.test.ts
@@ -14,6 +14,7 @@ import {
   jsonSchemaCredentialService,
   linkedVpOrg,
   linkedVpService,
+  mockDidDocumentChatbot,
 } from '../__mocks__'
 
 // --- Globals for test lifecycle ---
@@ -53,6 +54,28 @@ describe('Integration with Verana Blockchain', () => {
     fetchMocker.disable()
     vi.clearAllMocks()
   })
+
+  it('should perform a full integration self signed by resolving a real DID and validating the schema', async () => {
+    const did = 'did:web:dm.chatbot.demos.dev.2060.io'
+    const didResolverService = agent.dependencyManager.resolve(DidResolverService)
+    const didResolver = new Resolver({
+      web: async (did: string) => didResolverService.resolve(agentContext, did),
+    })
+
+    // Setup spy methods
+    const resolveSpy = vi.spyOn(Resolver.prototype, 'resolve')
+
+    const result = await resolve(did, {
+      didResolver,
+      agentContext,
+    })
+
+    // Validate result
+    expect(resolveSpy).toHaveBeenCalledTimes(1)
+    expect(resolveSpy).toHaveBeenCalledWith(did)
+    expect(result.verified).toEqual(true)
+    expect(JSON.parse(JSON.stringify(result.didDocument))).toEqual(mockDidDocumentChatbot)
+  }, 10000)
 
   it('should integrate with Verana testnet and retrieve the nested schema from the blockchain', async () => {
     const did = 'did:web:bcccdd780017.ngrok-free.app'
@@ -118,5 +141,5 @@ describe('Integration with Verana Blockchain', () => {
         },
       }),
     )
-  }, 20000)
+  }, 10000)
 })
